+++ server/lessons/RYRFcZXnLyJJm1Jc/evts.js
@@ -2,5 +2,8 @@ import * as a from "../../../src/lesson";
 return [
     new LE_AddGBubble([
         "Hello!"
-    ])
+    ],BubbleLoc.global,[
+        new AddFileTask("main.js",true),
+        new AddGenericCodeTask('console.log("hello");',"javascript")
+    ]),
 ]