+++ src/util.ts
@@ -310,7 +310,7 @@ class LogInMenu extends Menu{
         //     }
         // });
 
-        if(true){
+        if(false){
             this.body.innerHTML = `
             <div class="two-col">
                 <div class="g-cont">
@@ -347,23 +347,33 @@ class LogInMenu extends Menu{
                 this.body.textContent = "Failed to connect to the server.";
                 return;
             }
+            this.body.style.gap = "5px";
             this.body.innerHTML = `
-                <div class="two-col">
+                <div>Log in to access lessons, challenges, and an editor to experiment in!</div>
+                ${g_user ? `<div class="l-currently-logged-in">Currently logged in as ${g_user.email}</div>` : ""}
+                <div class="sign-in-btn"></div>
+                <!--<div class="two-col" style="grid-template-columns:unset">
                     <div>
                         <div class="sign-in-cont">
-                            <div>Sign in with Google</div>
+                            <div>Log in with Google</div>
                             <div class="si-icon-cont"><svg version="1.1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48" class="LgbsSe-Bz112c"><g><path fill="#EA4335" d="M24 9.5c3.54 0 6.71 1.22 9.21 3.6l6.85-6.85C35.9 2.38 30.47 0 24 0 14.62 0 6.51 5.38 2.56 13.22l7.98 6.19C12.43 13.72 17.74 9.5 24 9.5z"></path><path fill="#4285F4" d="M46.98 24.55c0-1.57-.15-3.09-.38-4.55H24v9.02h12.94c-.58 2.96-2.26 5.48-4.78 7.18l7.73 6c4.51-4.18 7.09-10.36 7.09-17.65z"></path><path fill="#FBBC05" d="M10.53 28.59c-.48-1.45-.76-2.99-.76-4.59s.27-3.14.76-4.59l-7.98-6.19C.92 16.46 0 20.12 0 24c0 3.88.92 7.54 2.56 10.78l7.97-6.19z"></path><path fill="#34A853" d="M24 48c6.48 0 11.93-2.13 15.89-5.81l-7.73-6c-2.15 1.45-4.92 2.3-8.16 2.3-6.26 0-11.57-4.22-13.47-9.91l-7.98 6.19C6.51 42.62 14.62 48 24 48z"></path><path fill="none" d="M0 0h48v48H0z"></path></g></svg></div>
                         </div>
                     </div>
                     <div>
-                        <!--<button>Continue</button>-->
+                        
                     </div>
-                </div>
+                </div>-->
             `;
-            let signInBtn = this.body.querySelector(".sign-in-cont") as HTMLElement;
-            signInBtn.addEventListener("click",e=>{
-                promptSignIn();
-            });
+            let signInBtn = this.body.querySelector(".sign-in-btn") as HTMLElement;
+            // let signInBtn = this.body.querySelector(".sign-in-cont") as HTMLElement;
+            // signInBtn.addEventListener("click",e=>{
+            //     promptSignIn();
+            // });
+            google.accounts.id.renderButton(signInBtn, {
+                theme: 'filled_blue', //outline
+                size: 'large',
+                text:"Log in with Google"
+            })
         }
 
         return this;
@@ -475,12 +485,8 @@ function logUserIn(data?:CredentialResData,token?:string){
         return;
     }
     // clear user - log out
-    h_profile.classList.remove("logged-in");
-    h_profile.innerHTML = `<span class="material-symbols-outlined">Person</span>`;
-    localStorage.removeItem("logData");
-    localStorage.removeItem("token");
-
-    _logout(data);
+    console.warn("... logging out?");
+    _logout();
 }
 
 // let testMenu = new SignUpMenu();
@@ -497,7 +503,7 @@ function genHeader(i:number,isCompact=true,id:string){
     if(id == "editor" || id == "lesson"){
         noLogo = true;
     }
-    let isLesson = (PAGE_ID == PAGEID.lesson);
+    let isLesson = (id == "lesson");
     navCont.className = "nav-container";
     navCont.role = "navigation";
     let logoI = 1;
@@ -517,15 +523,15 @@ function genHeader(i:number,isCompact=true,id:string){
             <!-- <span class="logo">Code Otter</span> -->
             
         </a>
-        </div>`:id=="editor"||id=="lesson"?`
+        </div>`:""}${id=="editor"||id=="lesson"?`
         <div class="editor-menu-bar">
-            <div class="b-editor-dashboard icon-div"><div class="material-symbols-outlined co-item" co-label="${isLesson ? "View Lessons" : "Project Dashboard"}">home</div></div>
+            <div class="b-editor-dashboard icon-div"><div class="material-symbols-outlined co-item" co-label="${isLesson ? "View Lessons" : "Project Dashboard"}">${isLesson ? "home" : "list"}</div></div>
             <!--<div>File</div>
             <div>Edit</div>
             <div>View</div>-->
         </div>
         <div class="cur-project-controls">
-            <div class="d-current-project">. . .</div>
+            <div class="d-current-project">[No Project Loaded]</div>
             <div class="b-save icon-div"><div class="material-symbols-outlined co-item" co-label="Save Project">save</div></div>
             <div class="b-publish icon-div hide"><div class="material-symbols-outlined"></div></div>
             <div class="icon-div hide"><div class="material-symbols-outlined">play_arrow</div></div>
@@ -2263,6 +2269,7 @@ function closeAllSubMenus(){
     for(const s of subs){
         s.e.remove();
         subs.splice(subs.indexOf(s),1);
+        if(s.ops.useHold) s.e.classList.remove("hold");
         // _hadClosedSubMenu = true;
         if(s.ops?.onclose) s.ops?.onclose();
     }
@@ -2273,12 +2280,13 @@ function areSubMenus(){
 }
 async function openDropdown(btn:HTMLElement,getLabels:()=>string[],onclick:(i:number)=>void,ops?:DropdownOptions){
     let labels = getLabels();
-    let icons = (ops?.getIcons ? ops?.getIcons() : null) as string[];
+    let icons = (ops?.getIcons ? ops?.getIcons() : []) as string[];
     // if(e.button != 2) return;
     // subMenus.innerHTML = "";
     if(ops.isRightClick) closeAllSubMenus();
     else{
         if(areSubMenus()){
+            btn.classList.remove("hold");
             closeAllSubMenus();
             return;
         }
@@ -2290,6 +2298,7 @@ async function openDropdown(btn:HTMLElement,getLabels:()=>string[],onclick:(i:nu
     // else await wait(1);
     // let rect = btn.getBoundingClientRect();
     let dd = document.createElement("div");
+    if(ops.useHold) btn.classList.add("hold");
     dd.oncontextmenu = function(e){
         e.preventDefault();
     };
@@ -2353,6 +2362,7 @@ type DropdownOptions = {
     customPos?:boolean;
     openToLeft?:boolean;
     isRightClick?:boolean;
+    useHold?:boolean;
 }
 function setupDropdown(btn:HTMLElement,getLabels:()=>string[],onclick:(i:number)=>void,ops?:DropdownOptions){
     btn.addEventListener("contextmenu",async e=>{
@@ -2470,6 +2480,12 @@ function whenEnter(elm:HTMLInputElement,f:(v:string)=>void){
     elm.addEventListener("blur",e=>{
         f(elm.value);
     });
+}
+
+let _hasAlertedNotLoggedIn = false;
+function alertNotLoggedIn(){
+    _hasAlertedNotLoggedIn = true;
+    new LogInMenu().load();
 }
 
 // Paul transfered stuff